feat: 完善精准计费及后台管理

+++ dto/task.go
@@ -38,7 +38,7 @@ type TaskDto struct {
 	UserId     int             `json:"user_id"`
 	Group      string          `json:"group"`
 	ChannelId  int             `json:"channel_id"`
-	Quota      int             `json:"quota"`
+	Quota      int64           `json:"quota"`
 	Action     string          `json:"action"`
 	Status     string          `json:"status"`
 	FailReason string          `json:"fail_reason"`
