+++ model/usedata_flow_test.go
@@ -119,7 +119,7 @@ func TestGetFlowQuotaDataUsesQuotaDataRoleSpecificDimensions(t *testing.T) {
 	require.Equal(t, "alice", adminRows[0].Username)
 	require.Equal(t, "vip", adminRows[0].UseGroup)
 	require.Equal(t, "east", adminRows[0].ChannelName)
-	require.Equal(t, 150, adminRows[0].Quota)
+	require.Equal(t, int64(150), adminRows[0].Quota)
 
 	selfRows, err := GetFlowQuotaData(900, 2000, "", 1, common.RoleCommonUser)
 	require.NoError(t, err)
@@ -129,7 +129,7 @@ func TestGetFlowQuotaDataUsesQuotaDataRoleSpecificDimensions(t *testing.T) {
 	require.Empty(t, selfRows[0].ChannelName)
 	require.Empty(t, selfRows[0].TokenName)
 	require.Equal(t, "vip", selfRows[0].UseGroup)
-	require.Equal(t, 175, selfRows[0].Quota)
+	require.Equal(t, int64(175), selfRows[0].Quota)
 }
 
 func TestLogQuotaDataSplitsRowsByUseGroupTokenChannelAndNode(t *testing.T) {
@@ -186,8 +186,8 @@ func TestLogQuotaDataSplitsRowsByUseGroupTokenChannelAndNode(t *testing.T) {
 	require.Equal(t, 1, rows[0].ChannelID)
 	require.Equal(t, "node-a", rows[0].NodeName)
 	require.Equal(t, 2, rows[0].Count)
-	require.Equal(t, 150, rows[0].Quota)
+	require.Equal(t, int64(150), rows[0].Quota)
 	require.Equal(t, 60, rows[0].TokenUsed)
 	require.Equal(t, "default", rows[1].UseGroup)
-	require.Equal(t, 25, rows[1].Quota)
+	require.Equal(t, int64(25), rows[1].Quota)
 }
