+++ relay/channel/task/taskcommon/helpers.go
@@ -92,6 +92,6 @@ func (BaseBilling) AdjustBillingOnSubmit(_ *relaycommon.RelayInfo, _ []byte) map
 }
 
 // AdjustBillingOnComplete returns 0 (keep pre-charged amount).
-func (BaseBilling) AdjustBillingOnComplete(_ *model.Task, _ *relaycommon.TaskInfo) int {
+func (BaseBilling) AdjustBillingOnComplete(_ *model.Task, _ *relaycommon.TaskInfo) int64 {
 	return 0
 }
