+++ relay/common/billing.go
@@ -7,7 +7,7 @@ import "github.com/gin-gonic/gin"
 type BillingSettler interface {
 	// Settle 根据实际消耗额度进行结算，计算 delta = actualQuota - preConsumedQuota，
 	// 同时调整资金来源（钱包/订阅）和令牌额度。
-	Settle(actualQuota int) error
+	Settle(actualQuota int64) error
 
 	// Refund 退还所有预扣费额度（资金来源 + 令牌），幂等安全。
 	// 通过 gopool 异步执行。如果已经结算或退款则不做任何操作。
@@ -17,8 +17,8 @@ type BillingSettler interface {
 	NeedsRefund() bool
 
 	// GetPreConsumedQuota 返回实际预扣的额度值（信任用户可能为 0）。
-	GetPreConsumedQuota() int
+	GetPreConsumedQuota() int64
 
 	// Reserve 将预扣额度补到目标值；若目标值不高于当前预扣额度则不做任何事。
-	Reserve(targetQuota int) error
+	Reserve(targetQuota int64) error
 }
